// src/app/models/ejecutivo-credito-reasignacion.ts
export class EjecutivoCreditoReasignacion {
  actualEjecutivo: string;
  actualEjecutivoId: number;
  actualEjecutivoUuid: string;
  anteriorEjecutivo: string;
  anteriorEjecutivoId: number;
  anteriorEjecutivoUuid: string;
  asignacionId: number;
  codMoneda: string;
  creditoId: number;
  desde: number;
  etapa: string;
  fechaFinal: string;
  fechaInicio: string;
  frecuencia: string;
  hasta: number;
  moneda: string;
  numCredito: string;
  originalAsignacionId: number;
  originalEjecutivo: string;
  originalEjecutivoId: number;
  originalEjecutivoUuid: string;
  reasignacionId: number;
  reasignar: boolean;
  socio: string;
  socioId: number;
  subEtapa: string;
  usuarioRegstro: string;
  fechaRegisto: number;
}